NetSec/requirements.txt
nfstream>=6.5.0
scapy>=2.5.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
scikit-learn>=1.3.0
tqdm>=4.65.0
einops>=0.6.0
